move remote tests to new remote testuite case
this follows the existing convention of having tests named the same as their module
